+++ pkg/gemini/client.go
@@ -9,7 +9,9 @@ import (
 
 // Client is a high-level client for Gemini to handle redirects and timeouts.
 type Client struct {
-	Timeout time.Duration
+	Timeout      time.Duration
+	MaxRetries   int
+	MaxRedirects int
 }
 
 // ClientResponse is a high-level client response.
@@ -17,6 +19,15 @@ type ClientResponse struct {
 	Response  *Response
 	Url       string
 	MimeTypes map[string]bool
+}
+
+// MakeClient makes the default client
+func MakeClient() Client {
+	return Client{
+		Timeout:      7 * time.Second,
+		MaxRetries:   3,
+		MaxRedirects: 5,
+	}
 }
 
 // NavigatePage gets the new url and page content pointed at by `url`.
@@ -28,7 +39,9 @@ func (c *Client) NavigatePage(rawurl string) (*ClientResponse, error) {
 	resChan := make(chan result, 1)
 
 	go func() {
-		for {
+		tries := 0
+		redirects := 0
+		for tries-1 < c.MaxRetries && redirects < c.MaxRedirects {
 			resp, err := Request(rawurl)
 			if err != nil {
 				resChan <- result{Err: err}
@@ -38,9 +51,12 @@ func (c *Client) NavigatePage(rawurl string) (*ClientResponse, error) {
 			case 1:
 				// 1X Input
 				resChan <- result{Response: resp}
+				tries++
+				return
 			case 2:
 				// 2X Success
 				resChan <- result{Response: resp}
+				tries++
 				return
 			case 3:
 				// 3X Redirect
@@ -54,20 +70,41 @@ func (c *Client) NavigatePage(rawurl string) (*ClientResponse, error) {
 					// Absolute url
 					rawurl = resp.Header.Meta
 				}
+				redirects++
 			case 4:
 				// 4X Temporary Failure
 				resChan <- result{Response: resp}
+				tries++
 			case 5:
 				// 5X Permanent Failure
 				resChan <- result{Response: resp}
+				tries++
+				return
 			case 6:
 				// 6X Client Certificate Required
 				resChan <- result{Response: resp}
+				tries++
+				return
 			default:
 				// Unrecognized status code
 				resChan <- result{
 					Err: errors.New("unrecognized status code"),
 				}
+				tries++
+				return
+			}
+
+			// Sleep before retrying
+			time.Sleep(100 * time.Millisecond)
+		}
+
+		if tries >= c.MaxRetries {
+			resChan <- result{
+				Err: errors.New("exceeded maximum number of retries"),
+			}
+		} else if redirects > -c.MaxRedirects {
+			resChan <- result{
+				Err: errors.New("exceeded maximum number of redirects"),
 			}
 		}
 	}()